% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_functions.R
\name{ANOVA_assumptions}
\alias{ANOVA_assumptions}
\title{One-way ANOVA Assumptions}
\usage{
ANOVA_assumptions(data, continuous, grouping)
}
\arguments{
\item{data}{Data frame or tibble.}

\item{continuous}{Unquoted column name for continuous outcome.}

\item{grouping}{Unquoted column name for grouping variable (factor).}
}
\description{
One-way ANOVA Assumptions
}
