% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_functions.R
\name{ANOVA_table}
\alias{ANOVA_table}
\title{One-way ANOVA table}
\usage{
ANOVA_table(data, continuous, grouping)
}
\arguments{
\item{data}{Data frame or tibble.}

\item{continuous}{Unquoted column name for continuous outcome.}

\item{grouping}{Unquoted column name for grouping variable (factor).}
}
\description{
One-way ANOVA table
}
