% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_qq.R
\name{correlation_qq}
\alias{correlation_qq}
\title{QQ Plots for Correlation Assessment}
\usage{
correlation_qq(data, x, y)
}
\arguments{
\item{data}{A data frame or tibble.}

\item{x}{Unquoted column name of the first numeric variable.}

\item{y}{Unquoted column name of the second numeric variable.}
}
\value{
A \code{ggplot} object.
}
\description{
Produces side-by-side QQ plots for two numeric variables to
  visually assess normality.
}
