% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_functions.R
\name{dependent_mean_median}
\alias{dependent_mean_median}
\title{Paired mean/median summary table}
\usage{
dependent_mean_median(data, col1, col2, accuracy = 3)
}
\arguments{
\item{data}{Data frame or tibble.}

\item{col1}{Unquoted column name for "before" values.}

\item{col2}{Unquoted column name for "after" values.}

\item{accuracy}{Integer rounding digits for summary statistics (default 1).}
}
\value{
A tibble summarizing mean (SD) and median (IQR) for before, after, and their differences.
}
\description{
Paired mean/median summary table
}
