% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_functions.R
\name{dependent_qq_plot}
\alias{dependent_qq_plot}
\title{dependent_qq_plot}
\usage{
dependent_qq_plot(data, var1, var2)
}
\arguments{
\item{data}{A data frame in wide format.}

\item{var1}{A string specifying the column name of the first measurement.}

\item{var2}{A string specifying the column name of the second measurement.}
}
\value{
A ggplot object representing the QQ plot of the paired differences.
}
\description{
Constructs a QQ plot for the difference scores between two paired measurements.
}
\examples{
dependent_qq_plot(a1c_measurements, "first_measurement", "second_measurement")
}
