% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_functions.R
\name{independence_test}
\alias{independence_test}
\title{Chi-Square Test of Independence}
\usage{
independence_test(data, var1, var2, alpha = 0.05)
}
\arguments{
\item{data}{A data frame or tibble}

\item{var1}{First unquoted categorical variable}

\item{var2}{Second unquoted categorical variable}

\item{alpha}{Significance level (default = 0.05)}
}
\value{
Printed test results
}
\description{
Chi-Square Test of Independence
}
\examples{
independence_test(gss_cat, marital, religion)
}
