% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_functions.R
\name{kruskal_HT}
\alias{kruskal_HT}
\title{Kruskal–Wallis Hypothesis Test}
\usage{
kruskal_HT(data, continuous, grouping, alpha = 0.05)
}
\arguments{
\item{data}{A data frame}

\item{continuous}{Numeric/continuous response variable.}

\item{grouping}{Factor or grouping variable (≥ 2 levels).}

\item{alpha}{Significance level for conclusion helper.}
}
\description{
Kruskal–Wallis Hypothesis Test
}
