% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_functions.R
\name{one_way_ANOVA}
\alias{one_way_ANOVA}
\title{One-way ANOVA hypothesis test}
\usage{
one_way_ANOVA(data, continuous, grouping, alpha = 0.05)
}
\arguments{
\item{data}{Data frame or tibble.}

\item{continuous}{Unquoted column name for continuous outcome.}

\item{grouping}{Unquoted column name for grouping variable (factor).}

\item{alpha}{Numeric significance level (default 0.05).}
}
\description{
One-way ANOVA hypothesis test
}
