% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_functions.R
\name{posthoc_dunn}
\alias{posthoc_dunn}
\title{Dunn's posthoc test (nonparametric)}
\usage{
posthoc_dunn(data, continuous, grouping, adjust = TRUE)
}
\arguments{
\item{data}{A data frame}

\item{continuous}{Numeric/continuous response variable.}

\item{grouping}{Factor or grouping variable (≥ 2 levels).}
}
\description{
Dunn's posthoc test (nonparametric)
}
