% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_functions.R
\name{posthoc_fisher}
\alias{posthoc_fisher}
\title{Fisher's test - one-way ANOVA}
\usage{
posthoc_fisher(data, continuous, grouping)
}
\arguments{
\item{data}{Data frame or tibble.}

\item{continuous}{Unquoted column name for continuous outcome.}

\item{grouping}{Unquoted column name for grouping variable (factor).}
}
\description{
Fisher's test - one-way ANOVA
}
