% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_functions.R
\name{posthoc_tukey}
\alias{posthoc_tukey}
\title{Tukey's test - one-way ANOVA}
\usage{
posthoc_tukey(data, continuous, grouping)
}
\arguments{
\item{data}{Data frame or tibble.}

\item{continuous}{Unquoted column name for continuous outcome.}

\item{grouping}{Unquoted column name for grouping variable (factor).}
}
\description{
Tukey's test - one-way ANOVA
}
