% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_functions.R
\name{two_prop_HT}
\alias{two_prop_HT}
\title{Two-Sample Hypothesis Test for Proportions}
\usage{
two_prop_HT(
  data,
  binary,
  grouping,
  event,
  p = 0,
  alternative = "two.sided",
  alpha = 0.05
)
}
\arguments{
\item{data}{A data frame or tibble}

\item{grouping}{Unquoted grouping variable with exactly 2 levels}

\item{alternative}{Type of test: "two.sided", "less", or "greater"}

\item{alpha}{Significance level (default = 0.05)}

\item{outcome}{Unquoted binary outcome variable (e.g., "yes"/"no", 1/0)}

\item{success}{The value of `outcome` considered a "success"}
}
\value{
Prints test result
}
\description{
Two-Sample Hypothesis Test for Proportions
}
\examples{
two_prop_HT(data = penguins, outcome = sex, group = species, success = "female")
}
