% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_functions.R
\name{two_way_ANOVA}
\alias{two_way_ANOVA}
\title{Two-way ANOVA hypothesis test}
\usage{
two_way_ANOVA(data, continuous, A, B, interaction = TRUE, alpha = 0.05)
}
\arguments{
\item{data}{Data frame or tibble.}

\item{continuous}{Unquoted column name for continuous outcome.}

\item{A}{Factor A}

\item{B}{Factor B}

\item{interaction}{logical statement to indicate if an interaction term should be included}

\item{alpha}{Numeric significance level (default 0.05).}
}
\description{
Two-way ANOVA hypothesis test
}
