% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_functions.R
\name{two_way_ANOVA_table}
\alias{two_way_ANOVA_table}
\title{Two-way ANOVA table}
\usage{
two_way_ANOVA_table(data, continuous, A, B, interaction = TRUE)
}
\arguments{
\item{data}{Data frame or tibble.}

\item{continuous}{Unquoted column name for continuous outcome.}

\item{A}{Factor A}

\item{B}{Factor B}

\item{interaction}{logical statement to indicate if an interaction term should be included}
}
\description{
Two-way ANOVA table
}
