% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_functions.R
\name{alt_symbol}
\alias{alt_symbol}
\title{Get Symbol Corresponding to Alternative Hypothesis}
\usage{
alt_symbol(alt)
}
\arguments{
\item{alt}{Character string specifying alternative hypothesis. One of "less", "greater", or "two.sided".}
}
\value{
A character symbol ("<", ">", or "!=") representing the alternative.
}
\description{
Get Symbol Corresponding to Alternative Hypothesis
}
