% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_functions.R
\name{conclusion}
\alias{conclusion}
\title{Print Conclusion Based on p-value and Significance Level}
\usage{
conclusion(p_val, alpha)
}
\arguments{
\item{p_val}{Numeric p-value.}

\item{alpha}{Numeric significance level.}
}
\value{
None. Prints conclusion to console.
}
\description{
Print Conclusion Based on p-value and Significance Level
}
