% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_functions.R
\name{cooks}
\alias{cooks}
\title{Cook's Distance Plot}
\usage{
cooks(
  fitted.lm,
  label = TRUE,
  show.threshold = FALSE,
  threshold = "convention",
  scale.factor = 0.5,
  n_labels_desired = NULL,
  label_height_nudge = 0
)
}
\arguments{
\item{fitted.lm}{A linear model created via `glm`, `lm`, or `aov`.}

\item{label}{Logical. Whether to label outliers on the plot. Default is TRUE.}

\item{show.threshold}{Logical. Whether to display threshold lines. Default is FALSE.}

\item{threshold}{Threshold method for outlier detection. One of `"baseR"`, `"matlab"`, or `"convention"`.}

\item{scale.factor}{Numeric. Scales point size and line width. Default is 0.5.}

\item{n_labels_desired}{Integer. Maximum number of outliers to label. Default is all.}

\item{label_height_nudge}{Numeric. Vertical offset for outlier labels. Default is 0.}
}
\value{
A `ggplot2` object.
}
\description{
Returns a Cook's distance plot for the supplied linear model. Outliers can be labeled,
and custom thresholding options are supported. Based on `gg_cooksd` from the `lindia` package.
}
\examples{
\dontrun{
model <- lm(bill_length_mm ~ bill_depth_mm + year, penguins)
cooks(model)
}
}
