% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_functions.R
\name{dependent_qq}
\alias{dependent_qq}
\title{dependent_qq}
\usage{
dependent_qq(data, col1, col2)
}
\arguments{
\item{data}{A data frame in wide format.}

\item{col1}{A string specifying the column name of the first measurement.}

\item{col2}{A string specifying the column name of the second measurement.}
}
\value{
A ggplot object representing the QQ plot of the paired differences.
}
\description{
Constructs a QQ plot for the difference scores between two paired measurements.
}
\examples{
dependent_qq_plot(a1c_measurements, "first_measurement", "second_measurement")
}
