% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_functions.R
\name{goodness_of_fit}
\alias{goodness_of_fit}
\title{Chi-Square Goodness of Fit Test}
\usage{
goodness_of_fit(data, categorical, expected = NULL, alpha = 0.05)
}
\arguments{
\item{data}{A data frame or tibble}

\item{categorical}{Unquoted categorical variable}

\item{expected}{A named vector of expected proportions (should sum to 1). If NULL, equal proportions are assumed.}

\item{alpha}{Significance level (default = 0.05)}
}
\value{
Printed test result
}
\description{
Chi-Square Goodness of Fit Test
}
