% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_functions.R
\name{normality_check}
\alias{normality_check}
\title{Diagnostic Plots: Q-Q Plot and Residuals Histogram}
\usage{
normality_check(
  model = NULL,
  data = NULL,
  continuous = NULL,
  function_of = NULL
)
}
\arguments{
\item{model}{Either a fitted regression model object of class \code{lm}, \code{glm}, or \code{aov}, 
or \code{NULL} if using \code{data}, \code{continuous}, and \code{function_of}.}

\item{data}{(Optional) A data frame containing the data.}

\item{continuous}{(Optional) The numeric predictor or outcome column in \code{data}.}

\item{function_of}{(Optional) The grouping or explanatory column in \code{data}.}
}
\value{
A \code{ggpubr} object containing the Q-Q plot and histogram of residuals displayed side by side.
}
\description{
Creates a Q-Q plot of standardized residuals and a histogram of residuals for a fitted regression model object
             or directly from data and variables. The plots are displayed side by side.
}
\examples{
\dontrun{
# Using an lm() object
fit <- lm(bill_length_mm ~ bill_depth_mm, data = palmerpenguins::penguins)
qq_hist_plot(fit)

# Using data, continuous and function_of
qq_hist_plot(data = palmerpenguins::penguins, 
             continuous = bill_length_mm, function_of = bill_depth_mm)
}

}
