% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_functions.R
\name{null_symbol}
\alias{null_symbol}
\title{Get Symbol Corresponding to Null Hypothesis}
\usage{
null_symbol(alt)
}
\arguments{
\item{alt}{Character string specifying alternative hypothesis. One of "less", "greater", or "two.sided".}
}
\value{
A character symbol ("≥", "≤", or "=") representing the null.
}
\description{
Get Symbol Corresponding to Null Hypothesis
}
