% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_functions.R
\name{outlier_count}
\alias{outlier_count}
\title{Count Suspected Outliers}
\usage{
outlier_count(df, model)
}
\arguments{
\item{df}{Data frame used in the model.}

\item{model}{A linear model created via `glm`, `lm`, or `aov`.}
}
\value{
A tibble summarizing counts of suspected and non-suspected outliers.
}
\description{
Counts the number of suspected outliers based on standardized residuals.
}
\examples{
model <- lm(bill_length_mm ~ bill_depth_mm + year, penguins)
outlier_count(penguins, model)
}
