% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_functions.R
\name{plot_residuals}
\alias{plot_residuals}
\title{plot_residuals}
\usage{
plot_residuals(data, continuous, grouping)
}
\arguments{
\item{data}{A data frame in long format.}

\item{continuous}{A string specifying the column name of the outcome variable.}

\item{grouping}{A string specifying the column name of the grouping variable (should have exactly 2 levels).}
}
\value{
A ggplot2 object with a scatterplot.
}
\description{
Constructs a scatterplot of residuals by groups for use in determining equality of variance.
}
\examples{
plot_residuals(penguins, "bill_length_mm", "sex")
}
