% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_functions.R
\name{profile_plot}
\alias{profile_plot}
\title{Profile plot for two-way ANOVA}
\usage{
profile_plot(data, continuous, xaxis, lines)
}
\arguments{
\item{data}{Data frame or tibble.}

\item{continuous}{Unquoted column name for continuous outcome.}

\item{A}{Factor A}

\item{B}{Factor B}

\item{interaction}{logical statement to indicate if an interaction term should be included}

\item{alpha}{Numeric significance level (default 0.05).}
}
\description{
Profile plot for two-way ANOVA
}
