% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_functions.R
\name{significant_line}
\alias{significant_line}
\title{Test for Significant Regression Line}
\usage{
significant_line(data, outcome, function_of, family = gaussian(), alpha = 0.05)
}
\arguments{
\item{data}{A data frame or tibble}

\item{outcome}{Unquoted outcome variable (y)}

\item{function_of}{Unquoted formula RHS (e.g., x or x1 + x2)}

\item{family}{A glm family (default is gaussian)}

\item{alpha}{Significance level (default = 0.05)}
}
\value{
Printed output of LRT result
}
\description{
Test for Significant Regression Line
}
