% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_functions.R
\name{variance_check}
\alias{variance_check}
\title{Residuals vs Fitted Diagnostic Plot}
\usage{
variance_check(
  model = NULL,
  data = NULL,
  continuous = NULL,
  function_of = NULL
)
}
\arguments{
\item{model}{Either a fitted regression model object of class \code{lm}, \code{glm}, or \code{aov}, 
or \code{NULL} if using \code{data}, \code{continuous}, and \code{function_of}.}

\item{data}{(Optional) A data frame containing the data.}

\item{continuous}{(Optional) The numeric predictor or outcome column in \code{data}.}

\item{function_of}{(Optional) The grouping or explanatory column in \code{data}.}
}
\value{
A \code{ggplot2} object representing the scatterplot of residuals vs fitted values.
}
\description{
Creates a residuals vs fitted values scatterplot for a fitted regression model object or directly from data and variables.
}
\examples{
\dontrun{
# Using an lm() object
fit <- lm(bill_length_mm ~ bill_depth_mm, data = palmerpenguins::penguins)
resid_fitted_plot(fit)

# Using data, continuous and function_of
resid_fitted_plot(data = palmerpenguins::penguins, 
                  continuous = bill_length_mm, function_of = bill_depth_mm)
}

}
